email_processing_agent/requirements.txt
# requirements.txt for email_processing_agent
imaplib2
msal
requests
python-dotenv
beautifulsoup4
